refactor(ExceptionHandling): general cleanup. removed unnecessary doc
