--- backend/burning-okr/burning-okr-app/src/main/java/org/burningokr/controller/exceptions/ErrorInformation.java
@@ -1,25 +1,0 @@
-package org.burningokr.controller.exceptions;
-
-import java.util.UUID;
-
-public class ErrorInformation {
-
-  private String errorInformation;
-  private String errorId;
-
-  public ErrorInformation(String errorInformation) {
-    this.errorInformation = errorInformation;
-    this.errorId = UUID.randomUUID().toString(); // fürs nachschauen im Log
-  }
-
-  // region Getter
-  public String getErrorInformation() {
-    return errorInformation;
-  }
-
-  public String getErrorId() {
-    return errorId;
-  }
-  // endregion
-
-}
